test(semantic): rename invariant-only files away from migrated_ prefix
Per #2058 process change "No 'spirit of' migrations". Four files
under the migrated_ prefix didn't actually migrate specific e2e
tests — they pinned generic invariants:

- migrated_action_shapes.rs -> action_shapes.rs
  Action-verb shape pinning ("the unit tests of the editor's
  action alphabet"). No e2e source claimed in docstring.
- migrated_bulk.rs -> invariants_bulk.rs
  "Small claims pulled from across the e2e suite" — vague,
  no per-test source mapping.
- migrated_misc.rs -> invariants_misc.rs
  "Miscellaneous wave-3 migrations" — same vague pattern.
- migrated_layout_cursor.rs -> layout_cursor.rs
  Hardware-cursor LayoutScenarios. No e2e source.

Dropping the migrated_ prefix removes them from the migration-
conventions lint's worklist (they'd otherwise be flagged for
missing e2e source cite). They remain TRUSTWORTHY per the audit
— just not migrations.

This reduces the lint's worklist from 35 violations to 27.
