+++ crates/fresh-editor/tests/semantic/invariants_bulk.rs
