+++ crates/fresh-editor/tests/semantic/invariants_misc.rs
